Update TestApplication.cpp
I chose to use a struct to pass the values and mutex needed for locking, in order to get rid of global variables all together. There are certainly other options, for example using the Windows API. I figured this direction was fine for this application. Let me know if otherwise.
- Greg Ladd